% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscar.R
\name{buscar_cnpj}
\alias{buscar_cnpj}
\title{Busca um CNPJ no site da Receita Federal}
\usage{
buscar_cnpj(cnpj, output = "both", dir = ".")
}
\arguments{
\item{cnpj}{número do CNPJ, com ou sem os caracteres especiais.}

\item{output}{tipo de output: "df" retorna uma \code{tibble}, "html" salva um arquivo HTML, "both" retorna a \code{tibble} e salva um aquivo HTML.}

\item{dir}{pasta onde o arquivo html será salvo. Default diretório atual.}
}
\value{
Se \code{output} for "df" ou "both", retorna uma \code{tibble} com resultados após scraping.
}
\description{
Realiza uma busca de um CNPJ na Receita Federal e salva resultados em arquivo.
}

